Add background drawable for self-search product items
This commit introduces a new drawable resource `product_item_background_self_search.xml`.

This drawable defines a shape with a solid color (`@color/selfSearch`), rounded corners with a 12dp radius, and a 0.5dp stroke with the color `#E676583C`. It will be used as the background for product items originating from a self-search.