feat: Introduce NumberFormatter utility
This commit introduces a new `NumberFormatter` utility class with a static `formatAmount` method.

This method formats a `double` value into a `String`. If the number is a whole number (e.g., 10.0), it is formatted as an integer ("10"). If it has a fractional part (e.g., 10.5), it is returned as a string representation of the double ("10.5").